pattern match that covers all interpretable TSTs
In order to save you the time of looking into what we want to interpret,
I've thrown together a pattern match that covers all typed ASTs (aka TSTs)
that can be interpreted.

Note that some of the trees described in http://den.sh/quasiquotes.html#syntax-overview
are ignored here. That's because typechecking changes the shapes of the trees,
and the guide linked above is about ASTs in general, not just TSTs.