interprets applications
Even though Apply nodes in untyped ASTs can express partial applications
and sugared calls to `apply` methods, TSTs ensure that Apply(Apply(...))
chains map to full invocations of real methods.

As mentioned by Denys in https://github.com/scalareflect/interpreter/pull/5,
we can exploit that fact that process the entire chains of applications
at once, especially given that quasiquotes provide us with such a facility
via q"$expr[..$targs](...$argss)".

However, to my surprise, this doesn't actually simplify code. Firstly,
supporting only full method applications means that instead of just applications
we need to support method applications and constructor applications separately.
Secondly, this also means that we need to duplicate evaluation logic
of selection between evalSelect (field selections) and evalApply
(method selections). Finally, with multi-arglist evaluations we need
an additional eval method that takes List[List[Tree]], which is annoying
because its erasure clashes with the one that takes List[Tree] and because
the implementation of the former very much duplicates the implementation
of the latter without a good way to achieve code reuse (all in all, this
is a minor annoyance, but it is sort of a last straw).

This comes as a big surprise, because it actually provides some sort of
justification for having Select and especially New implemented as they are
implemented currently in scalac, which I always thought to be an awkward
encoding for the corresponding language concepts.