instantiate now takes a Type, not a ClassSymbol
Java does have reified generics in the form of arrays, so just knowing
a symbol of something we're going to instantiate isn't going to cut it.

(It's enough to have a symbol when we want to express just List, but
we need a type to express something parameterized with other types).